+++ src/adapters/adapters.types.ts
@@ -2,4 +2,5 @@ export const ADAPTER_TYPES = {
 	AnchorApiSdk: Symbol.for("AnchorApiSdk"),
 	GoogleMapsService: Symbol.for("GoogleMapsService"),
 	AwsRekognitionService: Symbol.for("AwsRekognitionService"),
+	AmazonSesService: Symbol.for("AmazonSesService"),
 }